+++ app/refinements/as_score.rb
@@ -4,6 +4,11 @@ module AsScore
       (self == 0) ? "-" : "%.2f" % [self]
     end
   end
+  refine Fixnum do
+    def as_score
+      (self == 0) ? "-" : "%.2f" % [self]
+    end
+  end
   refine NilClass do
     def as_score
       "-"