+++ db/schema.rb
@@ -41,6 +41,7 @@ ActiveRecord::Schema.define(version: 3) do
     t.integer "number"
     t.string "name"
     t.string "element_type"
+    t.integer "level"
     t.string "info"
     t.float "base_value"
     t.string "credit"